Rename tx into signed_mantle_tx.
